+++ MiMenu_Back/Data/DTOs/Voucher/VoucherGetByIdDto.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; }
         public string Type { get; set; }
         public int Discount { get; set; }
-        public int BuyMinimum { get; set; }
+        public decimal BuyMinimum { get; set; }
         public bool Visibility { get; set; }
         public string DueDate { get; set; }
         public string CreateDate { get; set; }
